Move server dependencies and an empty instance of it to their own files
