+++ src/server.ts
@@ -1,28 +1,6 @@
 import { STATUS_CODE, STATUS_TEXT } from "@std/http";
-import { filterIncludesKey, parseParameters } from "./filter.ts";
-import { PublicSSHKey } from "./load_config.ts";
-import { PGPKey } from "./load_config.ts";
-import { getPGPTarget, servePGPKey, servePGPKeyList } from "./serve_pgp.ts";
 import { getContentType } from "./content-types.ts";
-import { serveKeys } from "./serve-keys.ts";
-import { serveHome } from "./serve-home.ts";
-
-/**
- * The dependencies required by the server.
- * We use this to make it easier to mock the dependencies in tests.
- */
-export interface ServerDependencies {
-  instanceName: string;
-  filterIncludesKey: typeof filterIncludesKey;
-  parseParameters: typeof parseParameters;
-  serveHome: typeof serveHome;
-  serveKeys: typeof serveKeys;
-  getPGPTarget: typeof getPGPTarget;
-  servePGPKey: typeof servePGPKey;
-  servePGPKeyList: typeof servePGPKeyList;
-  sshKeys: PublicSSHKey[];
-  pgpKeys: PGPKey[];
-}
+import { ServerDependencies } from "./dependencies.ts";
 
 /**
  * Start a simple http server that listens on the provided port and provides authorized keys based on query string filter
