+++ src/test_helpers.ts
@@ -1,0 +1,18 @@
+import { ServerDependencies } from "./dependencies.ts";
+
+/**
+ * An empty set of dependencies.
+ * Intended for use as a base in tests.
+ */
+export const emptyDependencies: ServerDependencies = {
+  filterIncludesKey: () => false,
+  parseParameters: () => ({}),
+  serveHome: () => new Response(""),
+  serveKeys: () => new Response(""),
+  getPGPTarget: () => undefined,
+  servePGPKey: () => new Response(""),
+  servePGPKeyList: () => new Response(""),
+  sshKeys: [],
+  pgpKeys: [],
+  instanceName: "unit-tests",
+};
